refactor(static-image): extract shared types state and request handling
